requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
jinja2
httpx
psycopg2-binary
pydantic-settings
python-dotenv
apscheduler
fpdf2
fastapi-mail
a2wsgi
pymysql
